+++ train-a/src/app/app.routes.ts
@@ -5,7 +5,7 @@ import { authGuard } from './auth/guards/auth/auth.guard';
 import { SigninComponent } from './auth/pages/signin/signin.component';
 import { SignupComponent } from './auth/pages/signup/signup.component';
 import { CarriageAdminPageComponent } from './carriages/pages/carriage-admin-page/carriage-admin-page.component';
-import { OrdersComponent } from './orders/pages/orders.component';
+import { OrdersComponent } from './orders/pages/orders-page/orders.component';
 import { paths } from './shared/configs/paths';
 import { StationAdminPageComponent } from './stations/pages/station-admin-page/station-admin-page.component';
 import { roleGuard } from './user/guards/role/role.guard';
