--- train-a/src/app/orders/pages/orders-page/orders-page.component.html
@@ -1,1 +1,0 @@
-<p>orders-page works!</p>
