--- train-a/src/app/orders/pages/orders-page/orders-page.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-orders-page',
-  standalone: true,
-  imports: [],
-  templateUrl: './orders-page.component.html',
-  styleUrl: './orders-page.component.scss',
-})
-export class OrdersPageComponent {}
